refactor(components): organize components into feature-based folders
