+++ frontend/src/hooks/usePrologAPI.js
@@ -15,16 +15,16 @@ export default function usePrologApi() {
       setLoading(true);
       setError(null);
 
-      // 🟦 LOG 1: Respuestas crudas que llegan desde el frontend
-      console.log("📥 Respuestas originales que llegan a Prolog:", answers);
+      // Respuestas crudas que llegan desde el frontend
+      console.log(" Respuestas originales que llegan a Prolog:", answers);
 
-      // 1. Convertimos respuestas → predicados Prolog
+      // Convertimos respuestas → predicados Prolog
       const predicates = formatPrologPayload(answers);
 
-      // 🟪 LOG 2: Predicados generados para Prolog
-      console.log("📤 Predicados enviados a Prolog:", predicates);
+      // Predicados generados para Prolog
+      console.log(" Predicados enviados a Prolog:", predicates);
 
-      // 2. Llamada al backend Prolog
+      // Llamada al backend Prolog
       const response = await fetch("http://localhost:3001/prolog/recommend", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
@@ -37,12 +37,12 @@ export default function usePrologApi() {
 
       const data = await response.json();
 
-      // 3. Guardamos recomendaciones
+      // Guardamos recomendaciones
       setRecommendations(data.recommendations || []);
 
       return data.recommendations || [];
     } catch (err) {
-      console.error("❌ Error Prolog:", err);
+      console.error(" Error Prolog:", err);
       setError(err.message);
       return [];
     } finally {
